backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "python-socketio[client]>=5.16.1",
    "sio>=0.1.0",
    "uvicorn>=0.41.0",
]
